+++ test/1.2/test-middleware-swagger-ui.js
@@ -34,6 +34,7 @@ var assert = require('assert');
 var async = require('async');
 var helpers = require('../helpers');
 var path = require('path');
+var pkg = require('../../package.json');
 var request = require('supertest');
 
 var rlJson = _.cloneDeep(require('../../samples/1.2/resource-listing.json'));
@@ -139,7 +140,7 @@ describe('Swagger UI Middleware v1.2', function () {
         try {
           helpers.createServer([rlJson, [petJson, storeJson, userJson]], {
             swaggerUiOptions: {
-              swaggerUiDir: path.join(__dirname, '..', 'browser', 'test-bower.html')
+              swaggerUiDir: path.join(__dirname, '..', '..', 'package.json')
             }
           }, function () {
             done(new Error('Should not initialize with options.swaggerUiDir pointing to a file'));
@@ -156,19 +157,19 @@ describe('Swagger UI Middleware v1.2', function () {
         // as requested, that's all this test needs to do.
         helpers.createServer([rlJson, [petJson, storeJson, userJson]], {
           swaggerUiOptions: {
-            swaggerUiDir: path.join(__dirname, '..', 'browser')
+            swaggerUiDir: path.join(__dirname, '..', '..')
           }
         }, function (app) {
           request(app)
-            .get('/docs/test-bower.html')
+            .get('/docs/package.json')
             .expect(200)
-            .expect('content-type', 'text/html; charset=UTF-8')
+            .expect('content-type', 'application/json')
             .end(function (err, res) {
               if (err) {
                 return done(err);
               }
 
-              assert.ok(res.text.indexOf('<title>Mocha Test Runner (Swagger Tools Bower Build)</title>') > -1);
+              assert.deepEqual(JSON.parse(res.text), pkg);
 
               done();
             });
@@ -177,3 +178,13 @@ describe('Swagger UI Middleware v1.2', function () {
     });
   });
 });
+
+
+
+
+
+
+
+
+
+