+++ test/controllers/Users.js
@@ -26,6 +26,6 @@
 
 var response = module.exports.response = 'controllers/Users swagger-router OK';
 
-module.exports.getById = module.exports._getById = function getById (req, res, next) {
+module.exports.getById = module.exports._getById = function getById (req, res) {
   res.end(response);
 };